refactor(web): move VersionHistory to a shared components directory
The builder directory is slated for retirement with the form editor; the
version-history dialog is shared by both editors, so it moves out to
components/version-history unchanged.
